requirements.txt
coverage==4.5.1
ecdsa==0.13
pbkdf2==1.3
pycryptodomex==3.4.12
